requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.43.0

# Image generation
google-genai>=1.0.0

# Instagram
instagrapi>=2.1.0

# Scheduling
apscheduler>=3.10.0

# Trend analysis / Media processing
opencv-python>=4.9.0
librosa>=0.10.0
ffmpeg-python>=0.2.0
openai-whisper>=20231117
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
tenacity>=8.2.0
Pillow>=10.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.27.0
